requirements.txt
inference-sdk>=0.9.0
pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
tqdm>=4.62.0
scikit-image>=0.19.0
matplotlib>=3.5.0 